Move base app to home page
